Issue: Hi guys,
I've never seen this one before...

The map in this order is for 5227 addresses, but as soon as you push it to the page where you select any extra copies and design service, it jumps up to 5506 addresses.

Can you take a look? It's under Staples CPAM Sherry Pratt. I'll tell her we're looking into it.